[ref] objects into smaller files, renamed HittableList
Description
Since we are getting more objects and the file is getting full
